requirements-dev.txt
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.4.0
mypy>=1.10.0
django-stubs>=5.0.0
djangorestframework-stubs>=3.15.0
factory-boy>=3.3.0






